Add files via upload
The presentation starts with the importance of storytelling in data science and moves on to what Airbnb is and how it is used as an introduction. Then presents the project scope and aim and moves on to -mostly visual- process documenting exploratory data analysis, cleaning, and organization of the data and the dashboard. Finally, it finishes with some key insights about what could be improved.